ACP-L2-DEL-WRK-EXECUTE: remove dead Workers root Execute path
Delete the obsolete workers.Service.Execute operation, ExecuteRequest/
ExecuteResult contracts, and the private normalize/adapt/provider
execution branch reachable only from that operation (the entire
pkg/services/workers/internal/service package). Production wiring
never populated Root.execute, so Service.Execute always returned
ErrExecuteUnavailable in production. WorkstationExecutionService
(StartWorkstationPool/StopWorkstationPool/DispatchWorkstation/
CancelWorkstationDispatch) is unchanged and is now the sole
production execution surface on the Workers root.

Also removes the Execute-exclusive materialization path
(materializeExecuteResultForDispatch) in factory_runtime, which had
no production caller, and reconciles the package-target-manifest,
ownership-inventory, and package-structure-baseline docs for the
deleted internal/service package and CloneSafeDiagnostics symbol.
