+++ docs/internal/baselines/package-structure-baseline.json
@@ -1979,11 +1979,6 @@
     },
     {
       "rule": "service-root-exported-function",
-      "filePath": "pkg/services/workers/execute_contracts.go",
-      "target": "CloneSafeDiagnostics"
-    },
-    {
-      "rule": "service-root-exported-function",
       "filePath": "pkg/services/workers/execution_context.go",
       "target": "ResolveProjectID"
     },
