+++ docs/internal/packaged-service-structure/package-target-manifest.json
@@ -467,7 +467,6 @@
     "pkg/services/workers/internal",
     "pkg/services/workers/internal/diagnostics",
     "pkg/services/workers/internal/interface",
-    "pkg/services/workers/internal/service",
     "pkg/services/workers/internal/services/runners",
     "pkg/services/workers/internal/services/runners/inference",
     "pkg/services/workers/internal/services/runners/internal/inference",
@@ -2701,11 +2700,6 @@
     },
     {
       "packagePath": "pkg/services/workers/internal/interface",
-      "disposition": "move",
-      "destination": "workers/internal"
-    },
-    {
-      "packagePath": "pkg/services/workers/internal/service",
       "disposition": "move",
       "destination": "workers/internal"
     },
