+++ pkg/services/workers/internal/root_shim_test.go
@@ -10,7 +10,7 @@ import (
 func TestNewRootConstructsPublishedWorkersService(t *testing.T) {
 	t.Parallel()
 
-	root, err := NewRoot(&recordingRuntimeAssembly{}, workstationswire.NewService(), nil)
+	root, err := NewRoot(&recordingRuntimeAssembly{}, workstationswire.NewService())
 	if err != nil {
 		t.Fatalf("NewRoot() error = %v", err)
 	}
