+++ pkg/services/workers/wire/wire.go
@@ -8,19 +8,6 @@
 package wire
 
 import (
-	"context"
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/portpowered/infinite-you/pkg/platform/logging"
-	"github.com/portpowered/infinite-you/pkg/services/workers"
-
-	workersinternal "github.com/portpowered/infinite-you/pkg/services/workers/internal"
-	executeservice "github.com/portpowered/infinite-you/pkg/services/workers/internal/service"
-	"github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runners"
-	runnerswire "github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runners/wire"
-	runtimeassemblywire "github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runtime_assembly/wire"
 	"github.com/portpowered/infinite-you/pkg/services/workers/internal/services/workstations/executor/agentrun"
 	"github.com/portpowered/infinite-you/pkg/services/workers/internal/services/workstations/invocation"
 	workstationswire "github.com/portpowered/infinite-you/pkg/services/workers/internal/services/workstations/wire"
@@ -32,166 +19,6 @@ var (
 	NewWorktree             = worktree.New
 	NewPlatformGitCommander = worktree.NewPlatformGitCommander
 )
-
-// ExecuteOptions are optional process-scoped Execute edges. Omitted values use
-// inert defaults so construction remains side-effect free.
-type ExecuteOptions struct {
-	Observe  workers.ObservationSink
-	Logger   logging.Logger
-	Clock    func() time.Time
-	Worktree workers.FactoryWorktreePreparer
-}
-
-// NewService constructs an inert Workers root from construction ports. It
-// composes the accepted root through parent-private runtime_assembly,
-// workstations, and runners (agent/script/inference) owner construction
-// without publishing owner types on the returned peer surface. Missing required
-// construction ports fail with a deterministic construction error and a nil
-// service.
-func NewService(
-	agentDependencies runners.AgentDependencies,
-	scriptConfig runners.ScriptConfig,
-	scriptDependencies runners.ScriptDependencies,
-	inferenceConfig runners.InferenceConfig,
-	inferenceDependencies runners.InferenceDependencies,
-	options ...ExecuteOptions,
-) (workers.Service, error) {
-	if err := validateConstructionPorts(
-		agentDependencies,
-		scriptConfig,
-		scriptDependencies,
-		inferenceConfig,
-		inferenceDependencies,
-	); err != nil {
-		return nil, err
-	}
-	agentRegistry, err := runnerswire.NewAgentRegistry(agentDependencies)
-	if err != nil {
-		return nil, fmt.Errorf("construct Workers: %w", err)
-	}
-	scriptRegistry, err := runnerswire.NewScriptRegistry(scriptConfig, scriptDependencies)
-	if err != nil {
-		return nil, fmt.Errorf("construct Workers: %w", err)
-	}
-	inferenceRegistry, err := runnerswire.NewInferenceRegistry(inferenceConfig, inferenceDependencies)
-	if err != nil {
-		return nil, fmt.Errorf("construct Workers: %w", err)
-	}
-	runnerRegistry, err := combineRunnerRegistries(agentRegistry, scriptRegistry, inferenceRegistry)
-	if err != nil {
-		return nil, err
-	}
-	runtimeAssembly, err := runtimeassemblywire.NewService(runnerRegistry, defaultBindingAssembler)
-	if err != nil {
-		return nil, err
-	}
-	executeOptions := ExecuteOptions{}
-	if len(options) > 0 {
-		executeOptions = options[0]
-	}
-	executeService, err := executeservice.New(executeservice.Dependencies{
-		Runners:   runnerRegistry,
-		Providers: agentDependencies.Providers,
-		Observe:   executeOptions.Observe,
-		Logger:    executeOptions.Logger,
-		Clock:     executeOptions.Clock,
-		Worktree:  executeOptions.Worktree,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("construct Workers: %w", err)
-	}
-	return workersinternal.NewRoot(
-		runtimeAssembly,
-		workstationswire.NewService(executeOptions.Logger),
-		executeService,
-	)
-}
-
-func validateConstructionPorts(
-	agentDependencies runners.AgentDependencies,
-	scriptConfig runners.ScriptConfig,
-	scriptDependencies runners.ScriptDependencies,
-	inferenceConfig runners.InferenceConfig,
-	inferenceDependencies runners.InferenceDependencies,
-) error {
-	if agentDependencies.Providers == nil {
-		return fmt.Errorf("construct Workers: agent Providers service is required")
-	}
-	if agentDependencies.Publish == nil {
-		return fmt.Errorf("construct Workers: agent progress publisher is required")
-	}
-	if strings.TrimSpace(scriptConfig.Command) == "" {
-		return fmt.Errorf("construct Workers: script command is required")
-	}
-	if scriptDependencies.CommandRunner == nil {
-		return fmt.Errorf("construct Workers: script command runner is required")
-	}
-	if scriptDependencies.FactoryDocs == nil {
-		return fmt.Errorf("construct Workers: script Factory docs loader is required")
-	}
-	if scriptDependencies.Now == nil {
-		return fmt.Errorf("construct Workers: script clock is required")
-	}
-	if scriptDependencies.Publish == nil {
-		return fmt.Errorf("construct Workers: script progress publisher is required")
-	}
-	if scriptDependencies.Record == nil {
-		return fmt.Errorf("construct Workers: script event recorder is required")
-	}
-	if strings.TrimSpace(inferenceConfig.Worker.Name) == "" {
-		return fmt.Errorf("construct Workers: inference worker name is required")
-	}
-	if inferenceDependencies.Models == nil {
-		return fmt.Errorf("construct Workers: inference Models service is required")
-	}
-	return nil
-}
-
-func combineRunnerRegistries(
-	agentRegistry runners.Service,
-	scriptRegistry runners.Service,
-	inferenceRegistry runners.Service,
-) (runners.Service, error) {
-	registrations := make([]runners.Registration, 0, 3)
-	for _, entry := range []struct {
-		service  runners.Service
-		identity string
-	}{
-		{agentRegistry, runners.AgentIdentity},
-		{scriptRegistry, runners.ScriptIdentity},
-		{inferenceRegistry, runners.InferenceIdentity},
-	} {
-		binding, err := entry.service.Resolve(runners.ResolutionRequest{
-			Identity: entry.identity,
-		})
-		if err != nil {
-			return nil, fmt.Errorf(
-				"construct Workers runner registry: resolve %s runner: %w",
-				entry.identity,
-				err,
-			)
-		}
-		registrations = append(registrations, runners.Registration{
-			Identity: binding.Identity,
-			Metadata: binding.Metadata,
-			Runner:   binding.Runner,
-		})
-	}
-	return runnerswire.NewService(registrations)
-}
-
-func defaultBindingAssembler(
-	_ context.Context,
-	role workers.RuntimeBuildRoleRequest,
-	_ workers.RuntimeBuildOpeningOptions,
-	selection workers.ResolvedRunnerSelection,
-) (workers.AssembledRuntimeBinding, error) {
-	return workers.AssembledRuntimeBinding{
-		RoleName:        role.Name,
-		RoleKind:        role.Kind,
-		RunnerSelection: selection,
-	}, nil
-}
 
 var NewFactoryDocsLoader = workstationswire.NewFactoryDocsLoader
